requirements.txt
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-huggingface>=0.0.1
huggingface_hub>=0.23.0
python-dotenv>=1.0.0
streamlit>=1.32.0
